staterecovery: improve test code reuse and reliability
