+++ coordinator/ethereum/test-utils/build.gradle
@@ -9,6 +9,7 @@ dependencies {
   implementation(project(":coordinator:core"))
   implementation(project(":coordinator:clients:smart-contract-client"))
   implementation(project(":jvm-libs:linea:web3j-extensions"))
+  implementation(testFixtures(project(":jvm-libs:linea:web3j-extensions")))
   implementation(project(":coordinator:ethereum:common"))
   implementation(project(":jvm-libs:linea:testing:file-system"))
   implementation("org.web3j:core:${libs.versions.web3j.get()}") {
