+++ jvm-libs/linea/testing/l1-blob-and-proof-submission/src/main/kotlin/net/consensys/linea/testing/submission/SubmissionTestHelper.kt
@@ -1,7 +1,11 @@
 package net.consensys.linea.testing.submission
 
+import linea.web3j.waitForTxReceipt
 import net.consensys.zkevm.coordinator.clients.smartcontract.LineaRollupSmartContractClient
 import net.consensys.zkevm.domain.Aggregation
+import org.web3j.protocol.Web3j
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.minutes
 
 /**
  * Submits blobs respecting aggregation boundaries
@@ -32,6 +36,7 @@ data class SubmissionTxHashes(
   val blobTxHashes: List<String>,
   val aggregationTxHashes: List<String>
 )
+
 fun submitBlobsAndAggregations(
   contractClient: LineaRollupSmartContractClient,
   aggregationsAndBlobs: List<AggregationAndBlobs>,
@@ -53,4 +58,30 @@ fun submitBlobsAndAggregations(
       ).get()
     }
     .let { SubmissionTxHashes(blobSubmissionTxHashes, it) }
+}
+
+fun submitBlobsAndAggregationsAndWaitExecution(
+  contractClient: LineaRollupSmartContractClient,
+  aggregationsAndBlobs: List<AggregationAndBlobs>,
+  blobChunksSize: Int = 6,
+  l1Web3jClient: Web3j,
+  waitTimeout: Duration = 2.minutes
+) {
+  val submissionTxHashes = submitBlobsAndAggregations(
+    contractClient = contractClient,
+    aggregationsAndBlobs = aggregationsAndBlobs,
+    blobChunksSize = blobChunksSize
+  )
+
+  l1Web3jClient.waitForTxReceipt(
+    txHash = submissionTxHashes.aggregationTxHashes.last(),
+    timeout = waitTimeout
+  ).also { txReceipt ->
+    if (txReceipt.status != "0x1") {
+      val lastAggregation = aggregationsAndBlobs.findLast { it.aggregation != null }!!.aggregation!!
+      throw IllegalStateException(
+        "latest finalization=${lastAggregation.intervalString()} failed on L1. receipt=$txReceipt"
+      )
+    }
+  }
 }
