+++ jvm-libs/linea/web3j-extensions/src/testFixtures/kotlin/linea/web3j/Pollers.kt
@@ -1,4 +1,4 @@
-package net.consensys.zkevm.ethereum
+package linea.web3j
 
 import org.web3j.protocol.Web3j
 import org.web3j.protocol.core.methods.response.TransactionReceipt
