+++ state-recover/test-cases/src/integrationTest/kotlin/linea/staterecover/StateRecoveryManualReplayToLocalStackIntTest.kt
@@ -14,12 +14,11 @@ import net.consensys.linea.jsonrpc.client.RequestRetryConfig
 import net.consensys.linea.jsonrpc.client.VertxHttpJsonRpcClientFactory
 import net.consensys.linea.testing.submission.AggregationAndBlobs
 import net.consensys.linea.testing.submission.loadBlobsAndAggregationsSortedAndGrouped
-import net.consensys.linea.testing.submission.submitBlobsAndAggregations
+import net.consensys.linea.testing.submission.submitBlobsAndAggregationsAndWaitExecution
 import net.consensys.toULong
 import net.consensys.zkevm.ethereum.ContractsManager
 import net.consensys.zkevm.ethereum.LineaRollupDeploymentResult
 import net.consensys.zkevm.ethereum.Web3jClientManager
-import net.consensys.zkevm.ethereum.waitForTxReceipt
 import org.apache.logging.log4j.Level
 import org.apache.logging.log4j.LogManager
 import org.assertj.core.api.Assertions.assertThat
@@ -107,28 +106,16 @@ class StateRecoveryManualReplayToLocalStackIntTest {
         }
       }
 
-    // wait until state recovery Besu and Shomei are up
-    val submissionTxHashes = submitBlobsAndAggregations(
-      contractClient = rollupDeploymentResult.rollupOperatorClient,
-      aggregationsAndBlobs = aggregationsAndBlobs,
-      blobChunksSize = 6
-    )
-
     val lastAggregationAndBlobs = aggregationsAndBlobs.findLast { it.aggregation != null }!!
     val lastAggregation = lastAggregationAndBlobs.aggregation!!
-
-    Web3jClientManager.l1Client.waitForTxReceipt(
-      txHash = submissionTxHashes.aggregationTxHashes.last(),
-      timeout = 2.minutes
-    ).also {
-      assertThat(it.status).isEqualTo("0x1")
-        .withFailMessage(
-          "finalization=${lastAggregation.intervalString()} tx failed! " +
-            "replay data is not consistent with L1 state, potential cause: " +
-            "data has L1 -> L2 anchoring messages and misses L1 Rolling Hash: tx=$it"
-        )
-      log.info("finalization={} executed on l1 tx={}", lastAggregation.intervalString(), it)
-    }
+    // wait until state recovery Besu and Shomei are up
+    submitBlobsAndAggregationsAndWaitExecution(
+      contractClient = rollupDeploymentResult.rollupOperatorClient,
+      aggregationsAndBlobs = aggregationsAndBlobs,
+      blobChunksSize = 6,
+      l1Web3jClient = Web3jClientManager.l1Client
+    )
+    log.info("finalization={} executed on l1", lastAggregation.intervalString())
 
     await()
       .untilAsserted {
